Guide

Implementation

Length and time are in units of wavelength and period. This normalization allows usage of relative permitivity and permeability in equations . Fields including electric, magnetic and current density are simply bundled as a vector of vectors of arrays . Boundary conditions pad the field arrays . PML paddings are multilayered, while All other boundaries add single layers. Paddings are stateful and permanent, increasing the size of field and geometry arrays. Finite differencing happens every update update and are coordinated to implictly implement a staggered Yee's grid .

Sources

If a source has fewer nonzero dimensions than the simulation domain, its signal will get normalized along its singleton dimensions. For example, all planar sources in 3d or line sources in 2d will get scaled up by a factor of 1/dx. This way, discretisation would not affect radiated power.

Main.Luminescent.PlaneWave — Type

function PlaneWave(f, dims; fields...)

Constructs plane wave source

Args

f: time function

dims: eg -1 for wave coming from -x face

fields: which fields to excite & their scaling constants (typically a current source, eg Jz=1)

Main.Luminescent.Source — Type

function Source(f, c, lb, ub, label=""; fields...)
function Source(f, c, L, label=""; fields...)

Constructs custom source. Can be used to specify uniform or modal sources

Args

f: time function

c: origin or center of source

lb: lower bounds wrt to c

ub: upper bounds wrt to c

L: source dimensions in [wavelengths]

fields: which fields to excite & their scaling constants (typically a current source, eg Jz=1)

Boundaries

Unspecified boundaries default to PML

Main.Luminescent.Periodic — Type

Periodic(dims)

periodic boundary

Main.Luminescent.PML — Type

function PML(dims, d=0.25f0, σ=20.0f0)

Constructs perfectly matched layers (PML aka ABC, RBC) boundary of depth `d` wavelengths 
Doesn't need to be explictly declared as all unspecified boundaries default to PML

Main.Luminescent.PEC — Type

PEC(dims)

perfect electrical conductor dims: eg -1 for -x side

Main.Luminescent.PMC — Type

PMC(dims)

perfect magnetic conductor

Monitors

Main.Luminescent.Monitor — Function

function Monitor(c, L; normal=nothing, label="")
function Monitor(c, lb, ub; normal=nothing, label="")

Constructs monitor which can span a point, line, surface, volume or point cloud monitoring fields or power.

Args

c: origin or center of monitor

L: physical dimensions of monitor

lb: lower bounds wrt to c

ub: upper bounds wrt to c

normal: flux monitor direction (eg normal to flux surface)

Main.Luminescent.power — Function

function power(m::MonitorInstance, u)

total power (Poynting flux) passing thru monitor surface

Main.Luminescent.flux — Function

function flux(m::MonitorInstance, u)

power density (avg Poynting flux) passing thru monitor

Physics

Main.Luminescent.update! — Function

function update!(u, p, t, field_padding, source_instances)
function update!(u1, u, p, t, field_padding, source_instances)

Updates fields for 3d. Please use update instead of update! when doing AD. Mutating update! Writes new fields either onto old fields or into buffer arrays u1

Main.Luminescent.update — Function

function update(u, p, t, field_padding, source_instances)

Updates fields for 3d in a manner amenable to AD. See also Mutating update!

GPU support

Simply use Flux.gpu to move simulation variables to GPU. This turns Arrays into CUDA arrays which get processed on GPU for both forward and backpropagation passes

Automatic differentiation adjoints

Compatible with Zygote.jl and Flux.jl. Please use update instead of update! when doing AD. See inverse design examples

Generative inverse design

Please contact us for latest scripts. We wrote Jello.jl, an innovative Fourier domain neural model to generate length scale controlled efficiently paramaterized geometry .

Comparison with other FDTD software

Our focus is on inverse design and topology optimization using adjoints from automatic differentiation. We love a streamlined API backed by a clear, concise and extensible codebase. Attention is paid to speed and malloc but that's not our main concern. Meep is the most popular open source FDTD package owing to its maturity and comprehensive features. It supports AD in certain cases using custom adjoints which we avoid in our package in favor of more flexibility . There are numerous commercial software eg Ansys Lumerical, Comsol and various EDA or SI software. TMK none of them supports AD for inverse design